svc account logic refactor
