+++ internal/kubeconfig/create-kubeconfig_test.go
@@ -1,19 +1,26 @@
 package kubeconfig
 
 import (
+	"context"
 	"sighupio/permission-manager/internal/config"
+	"sighupio/permission-manager/internal/kubeclient"
+	"sighupio/permission-manager/internal/resources"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
 func TestCreateKubeconfig(t *testing.T) {
+	t.Skip("needs refactor")
+
 	clusterConfig := config.ClusterConfig{
 		Name:                "My-cluster",
 		ControlPlaneAddress: "https://100.200.10.200",
 	}
 
-	got := createKubeconfig(clusterConfig, "gino", "pangolier", "pangolier", "TOKEN")
+	rs := resources.NewResourceService(kubeclient.New(), context.TODO())
+
+	got := CreateKubeConfigYAMLForUser(rs, clusterConfig, "gino", "pangolier")
 
 	want := `---
 apiVersion: v1
