+++ internal/resources/resources.go
@@ -26,7 +26,7 @@ type resourceService struct {
 
 // NewResourcesService returns a new instance of a ResourceService
 // allowing to interact with a K8s cluster via the given K8s client interface.
-func NewResourcesService(kc k8sclient.Interface, ctx context.Context) ResourceService {
+func NewResourceService(kc k8sclient.Interface, ctx context.Context) ResourceService {
 	return &resourceService{
 		kubeclient: kc,
 		context:    ctx,
